==> General/WhatsApp Mapping/Final 2/Step_2_Add_Lat_Long copy.ipynb <==
# Import some libraries
import numpy as np  # useful for many scientific computing in Python
import pandas as pd # primary data structure library

# Note:

# I manually added the country column to the data file generated
# by the previous notebook.
#
# Since there were only 26 unique countries, it wasn't that hard
# to do.
#
# I also had to manually add my own area code since the chat logs did
# not have a phone number listed for me, just my name.

# Load the data file
df = pd.read_csv("chat_area_codes_2.csv")
df.head()

# Check the data types
df.dtypes

# Load the world geo data
df_codes = pd.read_csv("world_country_and_usa_states_latitude_and_longitude_values.csv")
df.head()

# Load US geo data
df_area_codes = pd.read_csv("us-area-code-geo.csv")
df_area_codes.head()

# Load Canada geo data
df_ca_area_codes = pd.read_csv("ca-area-code-geo.csv")
df_ca_area_codes.head()

# Function to return latitude/longitude based on the country and/or area code of the row
def getLatitudeLongitude(row):
  # If the country is United States, search the US dataset for geo data
  if row['Country'] == 'United States':
    retVal = df_area_codes[['latitude', 'longitude']].loc[df_area_codes['area_code'] == row['Code']]
    # Did we find anything?
    if len(retVal) == 0:
      # No, then area code probably belongs to Canada.
      # Search that dataset instead
      retVal = df_ca_area_codes[['latitude', 'longitude']].loc[df_ca_area_codes['area_code'] == row['Code']]
  else:
    # Not US/Canada, so search world dataset
    retVal = df_codes[['latitude', 'longitude']].loc[df_codes['country'] == row['Country']]

  if len(retVal) == 1:                                        # Found exactly one row?  Should always be true
    return pd.Series([retVal.iloc[0, 0], retVal.iloc[0, 1]])  # Yes, return latitude/longitude
  return None                                                 # No, return None

# Test the function
ret = getLatitudeLongitude(df.iloc[0])
ret

# Create a copy of the original dataset
df2 = df.copy()

# Now, for each row, call our function and save the data to new columns
df2[['latitude', 'longitude']] = df2.apply(getLatitudeLongitude, axis=1)
df2.head()

# Verify that there are no null rows
df2[df2.isnull().any(axis=1)]

# Save the data to a CSV file
df2.to_csv("chat_geo_data.csv", index=False)
