==> General/WhatsApp Mapping/Final 2/Step_2_Add_Lat_Long.ipynb <==
# Import some libraries
import numpy as np              # useful for many scientific computing in Python
import pandas as pd             # primary data structure library
from bs4 import BeautifulSoup   # For webscraping
import requests

# Load the data file.
# Force Code to load as string type because we want to preserve the "+".
# This is because we will use this later to determine if the code is US/Canada
# area code or a country code.
df = pd.read_csv("chat_area_codes_from_members.csv", dtype={'Code': np.string_})
df.head()

# Check the data types
df.dtypes

# Load the country codes from the URL
countr_code_url = "https://countrycode.org/"
response = requests.get(countr_code_url)
soup = BeautifulSoup(response.content, "html.parser")
soup.title

rows = soup.find("table").find("tbody").find_all("tr")  # Extract all rows from the table on the page
country_codes = {}                                      # Initialize an empty dictionary to store the country + code
for row in rows:                                        # Loop through all the rows
    cells = row.find_all("td")                          # Extract the cells from the row
    country = cells[0].get_text()                       # First cell is the country
    code = cells[1].get_text()                          # Second cell is the calling code
    country_codes[code] = country                       # Save code/country to the dictionary

# Note: the calling code is the key and the country is the value.  This is
# because we want to look up the country by code later on.

# Display the first 5 elements in the dictionary to verify we did it right
first_few = dict(list(country_codes.items())[:5])
first_few

# Load Canada geo data
df_ca_area_codes = pd.read_csv("ca-area-code-geo.csv")
df_ca_area_codes.head()

# Function to return the country based on the calling code
def getCountry(code):
  if code[0] == "+":                                      # Does the code begin with "+"?
    new_code = code[1:]                                   # Yes, then extract the code without the "+"
    return country_codes[new_code]                        # Return the country
  elif int(code) in df_ca_area_codes['area_code'].values: # No "+", is the code in Canada?
    return 'Canada'                                       # Yes, return 'Canada'
  else:
    return 'United States'                                # No, default to "United States"

# Test the function with some values
test_code = '+49'
print('{} belongs to {}'.format(test_code, getCountry(test_code)))
test_code = '201'
print('{} belongs to {}'.format(test_code, getCountry(test_code)))
test_code = '204'
print('{} belongs to {}'.format(test_code, getCountry(test_code)))

df2 = df.copy()                                                         # Make a copy of the original data
df2['Country'] = df2.apply(lambda row: getCountry(row['Code']), axis=1) # Create country column based on the code

# Check to make sure there are no nulls
df2.isnull().sum()

# Check the value counts just to be curious
df2.value_counts()

# Load the world geo data
df_codes = pd.read_csv("world_country_and_usa_states_latitude_and_longitude_values.csv")
df_codes.head()

# Load US geo data
df_area_codes = pd.read_csv("us-area-code-geo.csv")
df_area_codes.head()

# Function to return latitude/longitude based on the country and/or area code of the row
def getLatitudeLongitude(row):
  ret_val = None
  match row['Country']:
    case 'United States':
      area_code = int(row['Code'])          # Covert the 'Code' from string to integer
      retVal = df_area_codes[['latitude', 'longitude']].loc[df_area_codes['area_code'] == area_code]
    case 'Canada':
      area_code = int(row['Code'])          # Covert the 'Code' from string to integer
      retVal = df_ca_area_codes[['latitude', 'longitude']].loc[df_ca_area_codes['area_code'] == area_code]
    case _:
      retVal = df_codes[['latitude', 'longitude']].loc[df_codes['country'] == row['Country']]

  if len(retVal) == 1:                                        # Found exactly one row?  Should always be true
    return pd.Series([retVal.iloc[0, 0], retVal.iloc[0, 1]])  # Yes, return latitude/longitude
  return None                                                 # No, return None

# Test the function with the first row of the dataset
row = df2.iloc[0]
ret = getLatitudeLongitude(row)
print('Code {} is located at {}, {}'.format(row['Code'], ret[0], ret[1]))

# Test the function with the last row of the dataset
row = df2.iloc[-1]
ret = getLatitudeLongitude(row)
print('Code {} is located at {}, {}'.format(row['Code'], ret[0], ret[1]))

# Create a copy of the dataset
df3 = df2.copy()

# Now, for each row, call our function and save the data to new columns
df3[['latitude', 'longitude']] = df3.apply(getLatitudeLongitude, axis=1)
df3.head()

# Verify that there are no null rows
df3[df3.isnull().any(axis=1)]

# Save the data to a CSV file
df3.to_csv("chat_geo_data.csv", index=False)
