==> General/WhatsApp Mapping/Final 2/Step_3_Map_chat_people.ipynb <==
import numpy as np  # useful for many scientific computing in Python
import pandas as pd # primary data structure library
import folium

print('Folium installed and imported!')

# Load the data
df = pd.read_csv("chat_geo_data.csv")
df.head()

# Define the world map centered around the US with a low zoom level
world_map = folium.Map(location=[37.0902, -95.7129], zoom_start=2)
#world_map

# instantiate a feature group for the chat people in the dataset
chat_people = folium.map.FeatureGroup()

# loop through the dataframe and add each row to the chat_people feature group
for lat, lng, in zip(df['latitude'], df['longitude']):
    chat_people.add_child(
        folium.vector_layers.CircleMarker(
            [lat, lng],
            radius=5, # define how big you want the circle markers to be
            color='yellow',
            fill=True,
            fill_color='blue',
            fill_opacity=0.6
        )
    )

# add chat_people to the world map and display it!
world_map.add_child(chat_people)
